+++ src/main/java/store/service/PurchaseService.java
@@ -4,13 +4,7 @@ import static store.constant.ExceptionMessage.PRODUCT_NOT_FOUND;
 import static store.constant.ExceptionMessage.PROMOTION_NOT_FOUND;
 
 import camp.nextstep.edu.missionutils.DateTimes;
-import java.util.List;
 import java.util.function.BiFunction;
-import java.util.function.Supplier;
-import store.builder.ReceiptBuilder;
-import store.dto.CalculateProductDto;
-import store.dto.CalculatePromotionDto;
-import store.dto.CalculateResultDto;
 import store.dto.PromotionDto;
 import store.dto.PurchaseDto;
 import store.entity.Product;
@@ -25,15 +19,13 @@ public class PurchaseService {
     private final ProductModel productModel;
     private final PromotionModel promotionModel;
     private final CartModel cartModel;
-    private final ReceiptBuilder receiptBuilder;
 
     public PurchaseService(ProductService productService, ProductModel productModel,
-                           PromotionModel promotionModel, CartModel cartModel, ReceiptBuilder receiptBuilder) {
+                           PromotionModel promotionModel, CartModel cartModel) {
         this.productService = productService;
         this.productModel = productModel;
         this.promotionModel = promotionModel;
         this.cartModel = cartModel;
-        this.receiptBuilder = receiptBuilder;
     }
 
     public boolean isApplicablePromotion(String productName) {
@@ -92,15 +84,6 @@ public class PurchaseService {
         purchaseProduct(purchaseDto.name(), product.getPrice(), purchaseDto.quantity(), 0);
     }
 
-    public void memberShip(Supplier<String> inputSupplier) {
-        String input = inputSupplier.get();
-        if (input.equals("Y")) {
-            cartModel.setMembership(true);
-            return;
-        }
-        cartModel.setMembership(false);
-    }
-
     public void purchaseProduct(String name, Integer price, Integer quantity, Integer freeCount) {
         if (freeCount == 0) {
             productModel.decreaseNormal(name, quantity);
@@ -109,14 +92,6 @@ public class PurchaseService {
         }
         productModel.decreasePromotion(name, quantity);
         cartModel.addPromotionQuantity(name, price, quantity, freeCount);
-    }
-
-    public String getCalculateResult() {
-        List<CalculateProductDto> productInfo = cartModel.calculate();
-        List<CalculatePromotionDto> promotionInfo = cartModel.calculatePromotion();
-        CalculateResultDto resultInfo = cartModel.result();
-        cartModel.clear();
-        return receiptBuilder.issue(productInfo, promotionInfo, resultInfo);
     }
 
     public String getStoredProductInfo() {
